baza #OMM-24 DONE Time 1h

+++ src/main/java/com/organizemymoney/ris/Models/Bancni_racun.java
@@ -2,7 +2,7 @@ package com.organizemymoney.ris.Models;
 
 public class Bancni_racun {
 
-	private int id;
-	private double stanje;
+	private Long id;
+	private int stanje;
 
 }
